Hooked up steering to the move_me goal.

+++ navigation/Awareness.h
@@ -155,7 +155,7 @@ public:
 	 * @param path The waypoints of the path will be stored here.
 	 * @return The number of waypoints in the path. 0 if no path could be found. A negative values means that something went wrong.
 	 */
-	int findPath(const WFMath::Point<3>& start, const WFMath::Point<3>& end, int radius, std::list<WFMath::Point<3>>& path) const;
+	int findPath(const WFMath::Point<3>& start, const WFMath::Point<3>& end, float radius, std::list<WFMath::Point<3>>& path) const;
 
 	/**
 	 * @brief Process the tile at the specified index.
